[TACHYON-1199] Extract newTestConf to super class
